Get ride of the src/types
