--- src/types/tilejson.ts
@@ -1,16 +1,0 @@
-export type TileJSON = {
-    tilejson: '2.2.0' | '2.1.0' | '2.0.1' | '2.0.0' | '1.0.0';
-    name?: string;
-    description?: string;
-    version?: string;
-    attribution?: string;
-    template?: string;
-    tiles: Array<string>;
-    grids?: Array<string>;
-    data?: Array<string>;
-    minzoom?: number;
-    maxzoom?: number;
-    bounds?: [number, number, number, number];
-    center?: [number, number, number];
-    vector_layers: [{id: string}]; // this is partial but enough for what we need
-};
