+++ test/bench/lib/tile_parser.ts
@@ -12,7 +12,7 @@ import {StyleLayerIndex} from '../../../src/style/style_layer_index';
 import type {StyleSpecification} from '@maplibre/maplibre-gl-style-spec';
 import type {WorkerTileResult} from '../../../src/source/worker_source';
 import type {OverscaledTileID} from '../../../src/source/tile_id';
-import type {TileJSON} from '../../../src/types/tilejson';
+import type {TileJSON} from '../../../src/util/util';
 import type {Map} from '../../../src/ui/map';
 import type {IActor} from '../../../src/util/actor';
 import {MessageType} from '../../../src/util/actor_messages';
